SISY-144: qsub job fix
